chore(db config): move db config into utils
